% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{alter.to.true.binary.multiple}
\alias{alter.to.true.binary.multiple}
\title{Take a two dimensional matrix with two unique values and change those values
to integers 0 and 1.}
\usage{
alter.to.true.binary.multiple(bin.chains)
}
\arguments{
\item{bin.chains}{A two dimensional vector with two unique elements.}
}
\description{
\code{alter.to.true.binary.multiple} takes a two dimensional matrix that
contains two unique values and switches all of the values with either integer
values 0 or 1, so that the resulting matrix has a one-to-one correspondence
with the original.
}
\details{
For this function to work properly, its argument should be checked first
using \code{\link{check.false.binary.multiple}} and only used in
\code{alter.to.true.binary} if \code{\link{check.false.binary.multiple}}
returns TRUE.
}
\examples{
alter.to.true.binary.multiple(matrix(data = c("A","B","A","B","A","B","B",
"A","A"), ncol = 3))
alter.to.true.binary.multiple(matrix(data = c(TRUE,TRUE,TRUE,FALSE,FALSE,
TRUE,FALSE,FALSE,TRUE,TRUE,FALSE), ncol = 3))
}

