% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{alter.to.true.binary}
\alias{alter.to.true.binary}
\title{Take a one dimensional vector with two unique values and change those values
to integers 0 and 1.}
\usage{
alter.to.true.binary(bin.chain)
}
\arguments{
\item{bin.chain}{A one dimesional vector with two unique elements.}
}
\description{
\code{alter.to.true.binary} takes a one dimensional vector that contains two
unique values and switches all of the values with either integer values 0 or
1, so that the resulting vector has a one-to-one corespondance with the
original.
}
\details{
For this function to work properly, its argument should be checked first
using \code{\link{check.false.binary}}, and only used in
\code{alter.to.true.binary} if \code{\link{check.false.binary}} returns TRUE.
}
\examples{
alter.to.true.binary(c("A","B","B","B","A","B","A","A"))
alter.to.true.binary(c(T,T,T,F,F,T,F,F,T,T,F))
}

