% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{alter_to_true_binary}
\alias{alter_to_true_binary}
\title{Take a one dimensional vector with two unique values and change those values
to integers 0 and 1.}
\usage{
alter_to_true_binary(bin_chain, success)
}
\arguments{
\item{bin_chain}{A one dimensional vector with two unique elements.}

\item{success}{Denotes the data entry to be counted for run statistics.}
}
\description{
\code{alter_to_true_binary} takes a one dimensional vector that contains two
unique values and switches all of the values with either integer values 0 or
1, so that the resulting vector has a one-to-one correspondence with the
original.
}
\details{
For this function to work properly, its argument should be checked first
using \code{\link{check_false_binary}}, and only used in
\code{alter_to_true_binary} if \code{\link{check_false_binary}} returns TRUE.
}
\examples{
alter_to_true_binary(c("A","B","B","B","A","B","A","A"))
alter_to_true_binary(c(TRUE,TRUE,TRUE,FALSE,FALSE,TRUE,FALSE,FALSE,TRUE,TRUE,
FALSE))
}

