% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{check.false.binary.multiple}
\alias{check.false.binary.multiple}
\title{Check if a two dimensional matrix has two unique elements.}
\usage{
check.false.binary.multiple(bin.chains)
}
\arguments{
\item{bin.chains}{A two dimensional matrix}
}
\description{
\code{check.false.binary.multiple} returns TRUE is there are two unique
elements in its argument, and returns FALSE if there are not two unique
elements in its argument.
}
\details{
This function checks every row and column element to see if they all share
the same two values.
}
\examples{
check.false.binary.multiple(matrix(data = c(1,0,1,0,1,0,0,1,1), ncol = 3))
check.false.binary.multiple(matrix(data = c("A","B","A","B","A","B","B","A",
"A"), ncol = 3))
check.false.binary.multiple(matrix(data = c(TRUE,TRUE,TRUE,FALSE,FALSE,TRUE,
FALSE,FALSE,TRUE,TRUE,FALSE), ncol = 3))
}

