% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{check.false.binary}
\alias{check.false.binary}
\title{Check if a vector has two unique elements.}
\usage{
check.false.binary(bin.chain)
}
\arguments{
\item{bin.chain}{A one dimension vector.}
}
\description{
\code{check.false.binary} returns TRUE if there are two unique elements in
its argument, and returns FALSE if there are not two unique elements in its
argument.
}
\details{
This function is not designed to be used outside of this package. It is
sufficiently simple as to be practically redundant to the user.
}
\examples{
check.false.binary(c(1,0,0,1,0,0,0,1))
check.false.binary(c("A","B","B","B","A","B","A","A"))
check.false.binary(c(TRUE,TRUE,TRUE,FALSE,FALSE,TRUE,FALSE,FALSE,TRUE,TRUE,
FALSE))
}

