% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{check.true.binary.multiple}
\alias{check.true.binary.multiple}
\title{Check if a two dimensional matrix has only integer elements 0 and 1.}
\usage{
check.true.binary.multiple(bin.chains)
}
\arguments{
\item{bin.chains}{A two dimensional matrix.}
}
\description{
\code{check.true.binary.multiple} returns TRUE if all of the elements in its
argument are integers 0 or 1, and FALSE if not all of the values in its
argument are integers 0 or 1.
}
\details{
This function checks every row and column element of its argument to see if
they contain either the integer values 0 or 1.
}
\examples{
check.true.binary.multiple(matrix(data = c(1,0,1,0,1,0,0,1,1), ncol = 3))
check.true.binary.multiple(matrix(data = c("A","B","A","B","A","B","B","A","A"), ncol = 3))
check.true.binary.multiple(matrix(data = c(F,T,F,T,F,T,T,F,F), ncol = 3))
}

