% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{check.true.binary}
\alias{check.true.binary}
\title{Check if a one dimensional vector has only integer elements 0 and 1.}
\usage{
check.true.binary(bin.chain)
}
\arguments{
\item{bin.chain}{A one dimensional vector.}
}
\description{
\code{check.true.binary} returns TRUE if all of the elements in the argument
are either integers of value 0 or 1. \code{check.true.binary} returns FALSE
if not all of the elements in the argument are integers of value 0 or 1.
}
\details{
This function checks every element of its argument to see if they contain
either the integer values 0 or 1.
}
\examples{
check.true.binary(c(1,0,0,1,0,0,0,1))
check.true.binary(c("A","B","B","B","A","B","A","A"))
check.true.binary(c(TRUE,TRUE,TRUE,FALSE,FALSE,TRUE,FALSE,FALSE,TRUE,TRUE,
FALSE))
}

