% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_swaps.R
\name{check_false_binary}
\alias{check_false_binary}
\title{Check if a vector has two unique elements.}
\usage{
check_false_binary(bin_chain)
}
\arguments{
\item{bin_chain}{A one dimensional vector.}
}
\description{
\code{check_false_binary} returns TRUE if there are two unique elements in
its argument, and returns FALSE if there are not two unique elements in its
argument.
}
\details{
This function is not designed to be used outside of this package. It is
sufficiently simple as to be practically redundant to the user.
}
\examples{
check_false_binary(c(1,0,0,1,0,0,0,1))
check_false_binary(c("A","B","B","B","A","B","A","A"))
check_false_binary(c(TRUE,TRUE,TRUE,FALSE,FALSE,TRUE,FALSE,FALSE,TRUE,TRUE,
FALSE))
}

