% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chi_sq_test_stat}
\alias{chi_sq_test_stat}
\title{Calculates the Pearson's chi square test statistic for a single binary chain
.}
\usage{
chi_sq_test_stat(bin_chain, n_chain_uniques)
}
\arguments{
\item{bin_chain}{A single binary chain of data in the form of an integer
vector.}

\item{n_chain_uniques}{A integer value representing the number of unique
values in \code{bin_chain}.}
}
\description{
\code{chi_sq_test_stat} takes a binary chain of data and calculates the
Pearson's chi square test statistic associated with it.
}

