% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chi_sq_test_stat_array}
\alias{chi_sq_test_stat_array}
\title{Calculate the chi square test statistics for many single binary chains.}
\usage{
chi_sq_test_stat_array(bin_chains, n_chain_uniques)
}
\arguments{
\item{bin_chains}{A integer matrix of binary chains of data, with each row
being a different chain.}

\item{n_chain_uniques}{A integer value representing the number of unique
values in \code{bin_chains}.}
}
\description{
\code{chi_sq_test_stat_array} takes a two dimensional matrix of many binary
chains of data and returns a numeric vector filled with a chi square test
statistic for each of them.
}

