% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chiSqTestStat}
\alias{chiSqTestStat}
\title{Calculates the Pearson's chi square test statistic for a single binary chain
.}
\usage{
chiSqTestStat(binChain, nChainUniques)
}
\arguments{
\item{binChain}{A single binary chain of data in the form of an integer
vector.}

\item{nChainUniques}{A integer value representing the number of unique
values in \code{binChain}.}
}
\description{
\code{chiSqTestStat} takes a binary chain of data and calculates the
Pearson's chi square test statistic associated with it.
}

