% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chiSqTestStatArray}
\alias{chiSqTestStatArray}
\title{Calculate the chi square test statistics for many single binary chains.}
\usage{
chiSqTestStatArray(binChains, nChainUniques)
}
\arguments{
\item{binChains}{A integer matrix of binary chains of data, with each row
being a different chain.}

\item{nChainUniques}{A integer value representing the number of unique
values in \code{binChains}.}
}
\description{
\code{chiSqTestStatArray} takes a two dimensional matrix of many binary
chains of data and returns a numeric vector filled with a chi square test
statistic for each of them.
}

