% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{i_dim_sum}
\alias{i_dim_sum}
\title{Sum all the values in the first dimension of a three dimensional integer
vector.}
\usage{
i_dim_sum(n, j, k)
}
\arguments{
\item{n}{A three dimensional integer vector.}

\item{j}{An integer that is a valid index of the second dimension of
\code{n}.}

\item{k}{An integer that is a valid index of the third dimension of
\code{n}.}
}
\description{
\code{i_dim_sum} takes a three dimensional integer vector, fixes the values
of the second and third dimensions, and then sums the values of all the
entries in the vector with those two values for its second and third
dimensions.
}

