% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ikDimSum}
\alias{ikDimSum}
\title{Sum all the values in the first and third dimensions of a three dimensional
integer vector.}
\usage{
ikDimSum(n, j)
}
\arguments{
\item{n}{A three dimensional integer vector.}

\item{j}{An integer that is a valid index of the second dimension of
\code{n}.}
}
\description{
\code{ikDimSum} takes a three dimensional integer vector, fixes the value
of the second dimension, and then sums the values of all the
entries in the vector with that value for its second dimension.
}

