% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indicate_run}
\alias{indicate_run}
\title{Indicate whether or not there is a run at a point in a integer vector.}
\usage{
indicate_run(bin_chain, p, i)
}
\arguments{
\item{bin_chain}{A binary chain of data in the form of an integer vector.}

\item{p}{A integer value representing the length of the run to test for.}

\item{i}{A integer value representing the location in \code{bin_chain} to
test for a run starting at.}
}
\description{
\code{indicate_run} takes an integer vector \code{bin_chain}, and two
integers, \code{p} and \code{i}, and tells the user if a run of length
\code{p} starting and index \code{i} in the form of a Boolean value.
}

