% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jk_dim_sum}
\alias{jk_dim_sum}
\title{Sum all the values in the second and third dimensions of a three dimensional
integer vector.}
\usage{
jk_dim_sum(n, i)
}
\arguments{
\item{n}{A three dimensional integer vector.}

\item{i}{An integer that is a valid index of the first dimension of
\code{n}.}
}
\description{
\code{jkDimSum} takes a three dimensional integer vector, fixes the value
of the first dimension, and then sums the values of all the
entries in the vector with that value for its first dimension.
}

