% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kDimSum}
\alias{kDimSum}
\title{Sum all the values in the third dimension of a three dimensional integer
vector.}
\usage{
kDimSum(n, i, j)
}
\arguments{
\item{n}{A three dimensional integer vector.}

\item{i}{An integer that is a valid index of the first dimension of
\code{n}.}

\item{j}{An integer that is a valid index of the second dimension of
\code{n}.}
}
\description{
\code{kDimSum} takes a three dimensional integer vector, fixes the values
of the first and second dimensions, and then sums the values of all the
entries in the vector with those two values for its first and second
dimensions.
}

