% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madras}
\alias{madras}
\title{Data from the Madras schizophrenia study}
\format{A data frame with 69 rows and 12 columns.}
\usage{
madras
}
\description{
A data set documenting the presence of thought disorders across 12 months
since hospitalization. Each column represents a single patient, and each
row represents a separate month. Months with thought disorders are encoded as
1 and those without are encoded as 0. Data is sourced from "Analysis of
Longitudinal Data" by Peter J. Diggle, Patrick J. Heagerty, Kung-Yee Liang,
and Scott L. Zeger. It was retrieved from
http://faculty.washington.edu/heagerty/Books/AnalysisLongitudinal/madras.data
. Individuals who did not have 12 months of post hospitalization data were
removed to make the data rectangular.
}
\keyword{datasets}

