% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metropolis}
\alias{metropolis}
\title{Generate independent data from a single binary chain.}
\usage{
metropolis(binChain, m, b)
}
\arguments{
\item{binChain}{A single binary chain of data represented by an integer
vector.}

\item{m}{An integer representing the number of swaps to be attempted.}

\item{b}{An integer representing the number of new chains of data to be
generated.}
}
\description{
\code{metropolis} takes a single binary chain of data in the form of an
integer vector and generates \code{b} new independent chains of data,
placing all of them in an integer matrix with the original data in the first
row.
}
\details{
\code{metropolis} works by taking the supplied \code{binChain}, and
attempting \code{m} swaps on it, only preforming a swap of elements if
doing so maintains the number of transitions between states in the resulting
chain. \code{metropolis} then takes the resulting chain, and attempts
\code{m} swaps on it again, then saving the resulting vector in a new row
of an output matrix. \code{metropolis} does this \code{b} times, each time
saving the resulting vector. Once all of the new data has been generated,
\code{metropolis} returns the newly built integer matrix, of which the first
row is the original chain of data \code{binChain}.
}

