% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiple_chi_sq_test_stat}
\alias{multiple_chi_sq_test_stat}
\title{Calculate the Pearson's chi square test statistic for a set of binary chains
of data.}
\usage{
multiple_chi_sq_test_stat(bin_chains, n_chain_uniques)
}
\arguments{
\item{bin_chains}{A two dimensional integer vector where each row is a
separate binary chain of data.}

\item{n_chain_uniques}{An integer value representing the number of unique
elements in the set of chains \code{bin_chains}.}
}
\description{
\code{multiple_chi_sq_test_stat} takes a two dimensional integer vector
\code{bin_chains} in which each row represents a single binary chain of data,
and calculates a Pearson's chi square test statistic for the entire set.
}

