% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiple_run_test_stat}
\alias{multiple_run_test_stat}
\title{Calculate the run test statistic for a set of binary chains of data and a
run of a certain length.}
\usage{
multiple_run_test_stat(bin_chains, p)
}
\arguments{
\item{bin_chains}{A two dimensional integer vector where each row is a
separate binary chain of data.}

\item{p}{An integer value representing the length of run to test for.}
}
\description{
\code{multiple_run_test_stat} takes a two dimensional integer vector
\code{bin_chains} in which each row represents a single binary chain of data,
and calculates a run test statistic for a run of length \code{p} for the
entire set.
}

