% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipleChiSqTestStat}
\alias{multipleChiSqTestStat}
\title{Calculate the Pearson's chi square test statistic for a set of binary chains
of data.}
\usage{
multipleChiSqTestStat(binChains, nChainUniques)
}
\arguments{
\item{binChains}{A two dimensional integer vector where each row is a
separate binary chain of data.}

\item{nChainUniques}{An integer value representing the number of unique
elements in the set of chains \code{binChains}.}
}
\description{
\code{multipleChiSqTestStat} takes a two dimensional integer vector
\code{binChains} in which each row represents a single binary chain of data,
and calculates a Pearson's chi square test statistic for the entire set.
}

