% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipleChiSqTestStatArray}
\alias{multipleChiSqTestStatArray}
\title{Calculate Pearson's chi square test statistics for many sets of binary
chains of data.}
\usage{
multipleChiSqTestStatArray(binChains, nChainUniques)
}
\arguments{
\item{binChains}{A three dimensional vector containing sets of chains of
binary data.}

\item{nChainUniques}{An integer value representing the number of unique
elements in the set of chains \code{binChains}.}
}
\description{
\code{multipleChiSqTestStatArray} takes a three dimensional vector
containing multiple sets of binary chains of data, and returns a numeric
vector with entries corresponding to the Pearson's chi square test
statistics of each set of binary chains of data.
}

