% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipleIndicateRun}
\alias{multipleIndicateRun}
\title{Indicate whether or not a run of a certain length exists starting at a
certain point.}
\usage{
multipleIndicateRun(binChain, p, i)
}
\arguments{
\item{binChain}{A one dimensional integer vector representing a binary chain
of data.}

\item{p}{An integer representing the length of run to test for.}

\item{i}{An integer representing a valid index of \code{binChain}.}
}
\description{
\code{multipleIndicateRun} takes a single binary chain \code{binChain}, a
valid index of that chain \code{i}, and a length of run \code{p} and tests
whether or not a run of that length starts at index \code{i}.
}

