% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipleRunTestStat}
\alias{multipleRunTestStat}
\title{Calculate the run test statistic for a set of binary chains of data and a
run of a certain length.}
\usage{
multipleRunTestStat(binChains, p)
}
\arguments{
\item{binChains}{A two dimensional integer vector where each row is a
seperate binary chain of data.}

\item{p}{An integer value representing the length of run to test for.}
}
\description{
\code{multipleRunTestStat} takes a two dimensional integer vector
\code{binChains} in which each row represents a single binary chain of data,
and calculates a run test statistic for a run of length \code{p} for the
entire set.
}

