% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multipleRunTestStatArray}
\alias{multipleRunTestStatArray}
\title{Calculate run test statistics for many sets of chains of binary data.}
\usage{
multipleRunTestStatArray(binChains, p)
}
\arguments{
\item{binChains}{A three dimensional integer vector containing sets of
chains of binary data.}

\item{p}{An integer representing the length of run to test for.}
}
\description{
\code{multipleRunTestStatArray} takes a three dimensional integer vector
containing multiple sets of binary chains of data, and returns a numeric
vector with entries corresponding to the run test statistics for runs of
length p for each set of binary chains of data.
}

