% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_counts_multiple}
\alias{n_counts_multiple}
\title{Second order transition counts for multiple binary chains.}
\usage{
n_counts_multiple(bin_chains, n_chain_uniques)
}
\arguments{
\item{bin_chains}{A two dimensional integer vector, each of whose rows
represents a single binary chain of data.}

\item{n_chain_uniques}{The number of unique values in the set of chains
\code{bin_chains}, represented as an integer value.}
}
\description{
\code{n_counts_multiple} counts the number of second order transitions in
a integer matrix whose rows represent individual binary chains. It returns
a three dimensional vector whose indices represent the type of transition,
and whose values represent the number of times that each transition occurs
in the set of chains.
}

