% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nCounts}
\alias{nCounts}
\title{Second order transition counts for a single binary chain.}
\usage{
nCounts(binChain, nChainUniques)
}
\arguments{
\item{binChain}{An integer vector representing a chain of data.}

\item{nChainUniques}{The number or unique values in the chain
\code{binChains}, represented as an integer value.}
}
\description{
\code{nCounts} counts the number of second order transitions in a binary
chain of data, then returns a three dimensional vector whose indicies
represent the type of transition, and whose values represent the number of
times that each transition occurs in the chain.
}

