% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nCountsMultiple}
\alias{nCountsMultiple}
\title{Second order transition counts for multiple binary chains.}
\usage{
nCountsMultiple(binChains, nChainUniques)
}
\arguments{
\item{binChains}{A two dimensional integer vector, each of whose rows
represents a single binary chain of data.}

\item{nChainUniques}{The number of unique values in the set of chains
\code{binChains}, represented as an integer value.}
}
\description{
\code{nCountsMultiple} counts the number of second order transitions in
a integer matrix whose rows represent individual binary chains. It returns
a three dimensional vector whose indices represent the type of transition,
and whose values represent the number of times that each transition occurs
in the set of chains.
}

