% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ok_tornado}
\alias{ok_tornado}
\title{Tornado data for Oklahoma, 1950 - 2015}
\format{A data frame with 66 rows and 366 columns.}
\usage{
ok_tornado
}
\description{
A data set documenting the presence of tornadic activity in Oklahoma. Each
row represents a year, starting with 1950, and ending in 2015, and each
column is a day of that year. Each day on which a tornado occurred is marked
encoded as a 1 and each day without tornadic activity is encoded as a 0. The
data is inspired by, but not drawn from, the paper "A Markov Chain Model of
Tornadic Activity" by Mathias Drton, Caren Marzban, Peter Guttorp, and Joseph
T. Schaefer. As mentioned in their paper, "day 366 of a non-leap year is
coded as a non-tornadic day. This is not expected to adversely affect the
results." Data is sourced from the National Weather Service Weather Forecast
Office in Norman, Oklahoma, and can be found at
http://www.srh.noaa.gov/oun/?n=tornadodata-ok-monthlyannual .
}
\keyword{datasets}

