% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.binary.test_methods.R
\name{plot.multiple.binary.test}
\alias{plot.multiple.binary.test}
\title{Produce plots from objects of class "multiple.binary.test".}
\usage{
\method{plot}{multiple.binary.test}(multiple.binary.test)
}
\arguments{
\item{multiple.binary.test}{An object of \code{\link{class}}
\code{multiple.binary.test}.}
}
\description{
Produces some interesting plots of data and test statistics of objects of
\link{class} \code{multiple.binary.test}.
}

