% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.binary.test_methods.R
\name{plot.single_binary_test}
\alias{plot.single_binary_test}
\title{Produce plots from objects of class "single_binary_test".}
\usage{
\method{plot}{single_binary_test}(x, ...)
}
\arguments{
\item{x}{An object of \code{\link{class}}
\code{single_binary_test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Produces some interesting plots of data and test statistics of objects of
\link{class} \code{single_binary_test}.
}

