% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.binary.test_methods.R
\name{print.multiple.binary.test}
\alias{print.multiple.binary.test}
\title{Print instructions for examining objects of class "multiple.binary.test".}
\usage{
\method{print}{multiple.binary.test}(x, ...)
}
\arguments{
\item{x}{An object of \code{\link{class}}
\code{multiple.binary.test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Instruct user to use \code{summary} and \code{print} to examine objects of
\link{class} \code{multiple.binary.test}.
}

