% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_binary_test_methods.R
\name{print.single_binary_test}
\alias{print.single_binary_test}
\title{Print instructions for examining objects of class "single_binary_test".}
\usage{
\method{print}{single_binary_test}(x, ...)
}
\arguments{
\item{x}{An object of \code{\link{class}}
\code{single_binary_test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Instruct user to use \code{summary} and \code{print} to examine objects of
\link{class} \code{single_binary_test}.
}

