% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_test_stat}
\alias{run_test_stat}
\title{Calculate the run test statistic for a single binary chain.}
\usage{
run_test_stat(bin_chain, p)
}
\arguments{
\item{bin_chain}{A binary chain of data in the form of an integer vector.}

\item{p}{An integer greater than one representing the length of run to test
for.}
}
\description{
\code{run_test_stat} takes an integer vector \code{bin_chain} of a chain of
binary data, and a integer \code{p} representing the length of run to test
for. It returns the run test stat for that chain of data.
}

