% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_test_stat_array}
\alias{run_test_stat_array}
\title{Calculate run test statistics for many binary chains.}
\usage{
run_test_stat_array(bin_chains, p)
}
\arguments{
\item{bin_chains}{A two dimensional integer matrix with each row denoting a
individual binary chain of data.}

\item{p}{An integer value representing the length of run to test for.}
}
\description{
\code{run_test_stat_array} takes an integer matrix with each row denoting a
binary chain of data and returns an integer vector with run test statistics
for runs of length \code{p} corresponding to each binary chain.
}

