% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rUnifInt}
\alias{rUnifInt}
\title{Generate a random indice of a vector}
\usage{
rUnifInt(binChain)
}
\arguments{
\item{binChain}{A single binary chain in the form of an integer vector.}
}
\description{
\code{rUnifInt} takes a one dimensional vector and randomly generates a
valid index of that vector from the discrete uniform distribution. It does
not return either the first or last indice.
}
\details{
\code{rUnifInt} works by taking the length of a supplied R integer vector,
subtracting two from it, and multiplying it by a random number from the
uniform distribution on the interval (0, 1). it then rounds this number up
to the nearest integer and returns it.
}

