% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runTestStat}
\alias{runTestStat}
\title{Calculate the run test statistic for a single binary chain.}
\usage{
runTestStat(binChain, p)
}
\arguments{
\item{binChain}{A binary chain of data in the form of an integer vector.}

\item{p}{An integer greater than one representing the length of run to test
for.}
}
\description{
\code{runTestStat} takes an integer vector \code{binChain} of a chain of
binary data, and a integer \code{p} representing the length of run to test
for. It returns the run test stat for that chain of data.
}

