% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snoqualmie}
\alias{snoqualmie}
\title{Rainfall data for the month of January in Snoqualmie Falls, 1948 - 1983.}
\format{A data frame with 36 rows and 31 columns.}
\usage{
snoqualmie
}
\description{
A dataset documenting the presence of precipitation in Snoqualmie Falls
Washington on every day of every January from 1948 to 1983. 1 indicates a day
with >=0.01 inches of precipitation, 0 indicates a day with <0.01 inches of
precipitation. Data sourced from Peter Guttorp's Stochastic Modeling of
Scientific Data, 1995.
}
\keyword{datasets}

