% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.binary.test_methods.R
\name{summary.multiple.binary.test}
\alias{summary.multiple.binary.test}
\title{Produce a summary of objects of class "multiple.binary.test".}
\usage{
\method{summary}{multiple.binary.test}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{class}}
\code{multiple.binary.test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints test statistics and p-values for objects of \link{class}
\code{multiple.binary.test}.
}

