% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_binary_test_methods.R
\name{summary.multiple_binary_test}
\alias{summary.multiple_binary_test}
\title{Produce a summary of objects of class "multiple_binary_test".}
\usage{
\method{summary}{multiple_binary_test}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{class}}
\code{multiple_binary_test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints test statistics and p-values for objects of \link{class}
\code{multiple_binary_test}.
}

