% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.binary.test_methods.R
\name{summary.single.binary.test}
\alias{summary.single.binary.test}
\title{Produce a summary of objects of class "single.binary.test".}
\usage{
\method{summary}{single.binary.test}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{class}}
\code{single.binary.test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints test statistics, p-values, and provides sample data for objects of
\link{class} \code{single.binary.test}.
}

