% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_binary_test_methods.R
\name{summary.single_binary_test}
\alias{summary.single_binary_test}
\title{Produce a summary of objects of class "single_binary_test".}
\usage{
\method{summary}{single_binary_test}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{class}}
\code{single_binary_test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints test statistics, p-values, and provides sample data for objects of
\link{class} \code{single_binary_test}.
}

