% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap}
\alias{swap}
\title{Swap elements of single binary chains}
\usage{
swap(binChain, m)
}
\arguments{
\item{binChain}{A binary one dimensional integer vector.}

\item{m}{A integer value representing the number of swaps to attempt.}
}
\description{
\code{swap} is used to swap elements of a single binary chain if doing so
maintains the same number of transitions between the two states of that
chain.
}
\details{
\code{swap} takes a one dimensional vector of integers \code{binChain} and
an integer \code{m}. It attempts to swap elements of \code{binChain}
\code{m} times, each time only completing the swap if it does not affect the
number of transitions between states in the sequence.
}

