% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap_mult}
\alias{swap_mult}
\title{Swap elements of multiple binary chains}
\usage{
swap_mult(bin_chains, m)
}
\arguments{
\item{bin_chains}{A two dimensional integer vector with binary values.}

\item{m}{A positive nonzero integer value for the attempted number of swaps
to attempt on \code{bin_chains}.}
}
\description{
\code{swap_mult} is used to swap elements of multiple binary chains if doing
so maintains the same number of transitions between the two states of those
chains.
}
\details{
\code{swap_mult} works by taking a two dimensional integer vector
\code{bin_chains} and \code{m}, a number of times to attempt swaps. It
generates random integers which are valid indices of the two dimensional
vector \code{bin_chains} and tries to swap the elements of the vector at
the indices that it generates, only doing so if this preserves the total
number of transitions between states. After attempting \code{m} swaps,
\code{swap_mult} returns the new, freshly swapped two dimensional vector of
binary chains.
}

