% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u1_test_stat}
\alias{u1_test_stat}
\title{Calculates the likelihood ratio test statistic for a single binary chain.}
\usage{
u1_test_stat(bin_chain, n_chain_uniques)
}
\arguments{
\item{bin_chain}{A binary chain of data in the form of a one dimensional
integer vector.}

\item{n_chain_uniques}{A integer value representing the number of unique
values in \code{bin_chain}.}
}
\description{
\code{u1_test_stat} takes a binary chain of data and calculates the likelihood
ratio test statistic associated with it.
}

