% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u1_test_stat_array}
\alias{u1_test_stat_array}
\title{Calculate likelihood ratio test statistics for many binary chains.}
\usage{
u1_test_stat_array(bin_chains, n_chain_uniques)
}
\arguments{
\item{bin_chains}{A two dimensional integer matrix with each row denoting a
individual binary chain of data.}

\item{n_chain_uniques}{An integer value representing the number of unique
values in the binary chains found in \code{bin_chains}.}
}
\description{
\code{u1_test_stat_array} takes an integer matrix with each row denoting a
binary chain of data and returns an integer vector with likelihood ratio test
statistics corresponding to each binary chain.
}

