% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u1TestStat}
\alias{u1TestStat}
\title{Calculates the liklihood ratio test statistic for a single binary chain.}
\usage{
u1TestStat(binChain, nChainUniques)
}
\arguments{
\item{binChain}{A binary chain of data in the form of a one dimensional
integer vector.}

\item{nChainUniques}{A integer value representing the number of unique
values in \code{binChain}.}
}
\description{
\code{u1TestStat} takes a binary chain of data and calculates the liklihood
ratio test statistic associated with it.
}

