% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u1TestStatArray}
\alias{u1TestStatArray}
\title{Calculate likelihood ratio test statistics for many binary chains.}
\usage{
u1TestStatArray(binChains, nChainUniques)
}
\arguments{
\item{binChains}{A two dimensional integer matrix with each row denoting a
individual binary chain of data.}

\item{nChainUniques}{An integer value representing the number of unique
values in the binary chains found in \code{binChains}.}
}
\description{
\code{u1TestStatArray} takes an integer matrix with each row denoting a
binary chain of data and returns an integer vector with likelihood ratio test
statistics corresponding to each binary chain.
}

