% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u6_test_stat_array}
\alias{u6_test_stat_array}
\title{Calculate likelihood ratio test statistics for many sets of binary chains of
data.}
\usage{
u6_test_stat_array(bin_chains, n_chain_uniques)
}
\arguments{
\item{bin_chains}{A three dimensional vector containing sets of chains of
binary data.}

\item{n_chain_uniques}{An integer value representing the number of unique
elements in the set of chains \code{bin_chains}.}
}
\description{
\code{u6_test_stat_array} takes a three dimensional vector containing multiple
sets of binary chains of data, and returns a numeric vector with entries
corresponding to the likelihood ratio test statistics of each set of binary
chains of data.
}

