% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{u6TestStatArray}
\alias{u6TestStatArray}
\title{Calculate liklihood ratio test statistics for many sets of binary chains of
data.}
\usage{
u6TestStatArray(binChains, nChainUniques)
}
\arguments{
\item{binChains}{A three dimensional vector containing sets of chains of
binary data.}

\item{nChainUniques}{An integer value representing the number of unique
elements in the set of chains \code{binChains}.}
}
\description{
\code{u6TestStatArray} takes a three dimensional vector containing multiple
sets of binary chains of data, and returns a numeric vector with entires
corresponding to the liklihood ratio test statistics of each set of binary
chains of data.
}

