% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec_greater_than}
\alias{vec_greater_than}
\title{Find the number of entries in a vector greater or equal to the value of the
first entry.}
\usage{
vec_greater_than(testStats)
}
\arguments{
\item{testStats}{A one dimensional numeric vector.}
}
\description{
\code{vec_greater_than} counts the number of entries in a numeric vector
\code{testStats} whose values are greater than that of the value of the
first element of the vector.
}

