% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vecGreaterThan}
\alias{vecGreaterThan}
\title{Find the number of entries in a vector greater or equal to the value of the
first entry.}
\usage{
vecGreaterThan(testStats)
}
\arguments{
\item{testStats}{A one dimensional numeric vector.}
}
\description{
\code{vecGreaterThan} counts the number of entries in a numberic vector
\code{testStats} whose values are greater than that of the value of the
first element of the vector.
}

