% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.binary.test_methods.R
\name{print.single.binary.test}
\alias{print.single.binary.test}
\title{Print instructions for examining objects of class "single.binary.test".}
\usage{
\method{print}{single.binary.test}(single.binary.test, ...)
}
\description{
Instruct user to use \code{summary} and \code{print} to examine objects of
\link{class} \code{single.binary.test}.
}

