% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.binary.test_methods.R
\name{summary.multiple.binary.test}
\alias{summary.multiple.binary.test}
\title{Produce a summary of objects of class "multiple.binary.test".}
\usage{
\method{summary}{multiple.binary.test}(multiple.binary.test, ..)
}
\description{
Prints test statistics and p-values for objects of \link{class}
\code{multiple.binary.test}.
}

