% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.binary.test_methods.R
\name{summary.single.binary.test}
\alias{summary.single.binary.test}
\title{Produce a summary of objects of class "single.binary.test".}
\usage{
\method{summary}{single.binary.test}(single.binary.test, ...)
}
\description{
Prints test statistics, p-values, and provides sample data for objects of
\link{class} \code{single.binary.test}.
}

