% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap}
\alias{swap}
\title{Swap elements of single binary chains}
\usage{
swap(binChain, m)
}
\arguments{
\item{binChain}{A binary one dimensional integer vector.}

\item{i}{An integer which is a valid indice of the vector binChain.}

\item{j}{An integer which is a valid indice of the vector binChain.}
}
\description{
\code{swap} is used to swap elements of a single binary chain if doing so
maintains the same number of transitions between the two states of that
chain.
}
\details{
\code{swap} takes a R integer vector, as well as two integers, i and j,
which are valid indicies of that vector. It then checks if swapping the
values at indicies i and j would change the number of transitions between
states in the vector, and if the swap does not, it switches the two values,
returning a new vector with the values switched. If it cannot switch the
values, it returns the original vector.
}

